Extense the configuration of consult.
Many settings are borrowed from consult wiki and from
https://blog.sumtypeofway.com/posts/emacs-config.html

Try to be consistent with function identification as for general-interface.el.

Suppress avy-flycheck and rubocop packages.
As well, rubocop is used from flycheck.

Improve the configuration of attrap.

Suppress lambda for hooks as far as possible. This is the case for
ruby-mode and python-mode thanks to my/add-hooks.

Add crystal-mode.
Suppress support for maxima, vcard, recutils, id-utils, pariemacs, .desktop
files and asymptote.

Fix configuration for latex-help, now it effectively works.