Add or change many bindings.
Unbind some useless command in help-map.

Bind calc to <XF86Calculator>.

Put consult-imenu in C-c map. (C-c m).

Bind ace-window to M-o

Add three bindings in helpful-mode-map
 - c to quickly customize `help-customise'
 - i to go to info `help-goto-info'
 - s to view source `help-view-source'

Option to show switches passed to ls in dired in the modeline
`dired-switches-in-mode-line'

Add a hook for view-mode to effectively be in emacs-state.

Activate global-auto-revert-mode for all buffers except for buffer listing.

Remove the package vdiff-magit