Several changes.
Add deadgrep-mode to be in initial evil-emacs-state.

Suppress the error for motion in macro recording and playing.

Add a new prefix `SPC T' in leader-ala-vim for toggling commands,
for now there are two toggling in this, whitespace-mode,
highlight-indent-guides and choose-folding-method.

Changes many declaration with general. Use :init instead of :general when
there are declarations for several maps/states.

Add actions to use ace-window from embark-act (very handy feature from:
https://karthinks.com/software/fifteen-ways-to-use-embark/).

Change bindings for embark. Use `C-c b' and `SPC b'. The latter is more
accessible and the former is consistent with the latter.
Add a binding to access embark-act from the minibuffer prompt: `C-;'

Reorganize the order of declarations.