Many improvements for avy.
Add a binding to avy-goto-char-time M-j
Add many commands to the menu dispatcher of avy :
 - kill/copy/yank (delete/yank/paste) the whole line.
 - call to embark-act
 - call to helpful-at-point
 - lookup dictionnary at point (using dictd as it's the default of
   the dictionnary builtin package).

Add the package avy-embark-collect.

Not relative to avy.
Bind M-y to consult-yank-pop, "SPC /" to consult-line, "SPC gc" to
consult-ripgrep.

Change the binding for marginalia-cycle from M-a to M-m.

Add the package visual-regexp (to try).
Add the package dumb-jump.

Set view-mode for buffer in read-only.