Move all files use by packages in var/.
Create a function to set several variables for this.
So, the .emacs.d is cleaner. There is nothing in .cache anymore.

The only issue is yasnippets. It insiste to create the directory
.emacs.d/snippets. I change the constant in the source, for now.
It should be not a constant but rather a custom variable, imho.
Anyway, yasnippets was not update since 2 years ago.