Several changes.
Replace (require 'dash) with (use-package dash) to bootstrap more
easily.

Remove another key binding in help-map (C-\).

Replace just-one-space by cycle-space (new function in emacs 29).

Suppress the package restart-emacs, as there is a command to restart
in emacs 29.

Improvements in my menu for toggling things and for xref.

Unbind C-M-h to be able to use which-key C-h.

Add many descriptions for prefix keys in which-key menus.

Add a use-package declaration for embark-consul.

Suppress the packages libgit and magit-libgit.