Several changes.
Modify the declaration for yasnippet. Declare the use of yasnippet-snippets.
Put a description for the prefix key `C-c &'.

Add settings for highlight-indent-guides. With emacs 29 the default values
doesn't render wery well.

Suppress the package hasky-stack.
Change the loading of hasky-extensions. Since it's no more on melpa, just
git clone the githup repo and use emacs autoloading.