Improve autoloading and key bindings.
Remove a key binding in help-map: h

Reinstall the key binding C-x r e to edit-bookmars. For a myterious
reason, this disapears, although it exists in bookmark-map.

Change the key bindings to function goto-last-change and
goto-last-reverse-change.
This is more appropriate to put these under "M-g" prefix (goto commands)
than under "M-s" (search commands).

Change key binding of expand-region from "SPC =" to "SPC $".
This is to put consult-flyspell on "SPC =", as in vim "z ="
permits to correct a word at point.

Defer loading of packages:
 - dabbrev
 - cape
 - symbol-overlay
 - consult-flyspell
 - consult-recoll
 - emojify
 - iedit

Improve declaration for all-the-icons-ibuffer, use ibuffer-hook is the
right way to get it to work.

Improve declaration of all-the-icons-completion. Move the :init to :config to
speed up the emacs startup.

For the package visual-replaceChange change key bindings for vr/replace from
"SPC ?" to "SPC !". Add key binding for vr/query-replace to "SPC ?".
Add a key binding to vr/query-replace: "C-c q".

Improve declaration for dumb-jump package. Add commands for autoloading.
So, now it works effectively.