Many changes.
Change the font in modeline to all items fit in screen.
Suppress #'kbd for consult and embark as they read string binding
format.
Add a hook for mail-mode to use flyspell minor mode.

Suppress the binding to #'consult-apropos, It's deprecated.
Add a binding to #'consult-info.

Remove the package emojify that I have never used.